fix `idlib_power_of_two_gt`
The implementation for `uint16_t` was declared using `uint8_t`.

+++ library/includes/idlib/bit_scan/power_of_two_gt.h
@@ -35,8 +35,8 @@ idlib_power_of_two_gt_u8
 int
 idlib_power_of_two_gt_u16
   (
-    uint8_t* RETURN,
-    uint8_t x
+    uint16_t* RETURN,
+    uint16_t x
   );
 
 int
